+++ koocook_core/controllers/__init__.py
@@ -1,2 +1,3 @@
 from .posts import PostHandler, PostController
 from .users import UserController, UserHandler
+from .decorators import to_koocook_user, apply_author_from_session
